epub-maker: extract buffered cover preview component
